Organize components in a folder and rename files to index.tsx
