+++ src/components/CarInfoPreview/CarInfoPreview.scss
