+++ src/components/Spiner/Spinner.scss
