+++ src/components/VinInput/VinInput.scss
